Update configuration handling
Instead of the 'dvctracker_settings.cfg' file, config settings will be
loaded from a python object. This allows for different configurations
depending on where DVCTracker is run.

The Heroku config setting adds the Werkzeug ProxyFix middleware to
maintain the information from the server that made the original request.

Also, .gitattributes is no longer needed as Github accurately labels the
project as Python.