Update base_parser object
Updated the base_parser object to allow for more configuration. The new
values that can be passed in are:
-Request Headers
-Request Parameters
-Parser Site URL
-Parser Data URL (For DVC Rental Store the data for the specials comes
from a different URL than the website itself. Passing two different URls
allows the special data to be accessed and the website URL to be stored
in the DB with the specials. The stored website URL is what will then be
presented to the user in the email.)