Add new DVC Rental Store parsers
The two parser modules for DVC Rental store are added in this commit. In
doing this some changes had to be made to accommodate parsing json data
instead of html. The ParsedSpecial object now has a 'raw_string'
property instead of a 'raw_strings' property. Rather than a list of the
strings that represent the special, one string is now used. Newlines in
this will be rendered in the error email because the raw text is now
wrapped in a <pre> tag in the template. The other change is the
'store-specials-data' command now takes an optional extension
parameter.

Some other changes in this commit:
-Till URL is now a value in the flask application object.
-Changed some SQLAlchemy queries to improve performance.
-Fixed some errors in how specials were tracked in 'update-specials'.
-Added some more character to the console output messages.
-Updated the Pipfile.lock and requirements.txt.