Add ParserStatus
Add the ParserStatus model to keep track of the health of a parser.
Parsers don't need to have a status but if 0 specials are found this
might indicate something is wrong. Rather than removing all of the
specials from the db it is better to keep everything as is and wait for
the user to decide if it is okay that the parser is empty. This can be
accomoplished by setting `empty_okay` to `True`.

Also, the update won't fail just because a single parser isn't working
anymore.