Add push and user cli commands and use APNS
Adds two convenience commands. The first creates a new user with a given
username and password. The second produces a base64 encoded string of
the p8 AuthKey file needed for APNS. The string can be stored as an
environment variable rather than having to worry about committing the p8
file into the repo. This also adds using push notifications when
updating specials and sending errors.