Separate rooms from views for specials
This separates special's rooms and views both in the db and in the
template. It also adds views as an important criteria.

This also slightly changes how some important criteria are found. Since
the resorts, rooms, and views will now be added based on what is on the
DVC site we can search for an exact match rather than doing a contains
check. This also replaces the lambdas with normal functions for sake of
clarity.