Add new models to the db
This commit adds a lot of new functionality to the db. With this commit
the following changes have been made:

- The 'type' for 'StoredSpecials' is now a Python Enum.

- Adds characteristic based models: Resort, Room, View.
  - These are 'Single Table Inheritance' models as they all derive from
    the Characteristic model (characteristics table in the db).
  - Doing this allows for simpler querying of all characteristics at
    one time.

- Adds category based models: ResortCategory, RoomCategory,
  ViewCategory.
  - These are 'Single Table Inheritance' models as they all derive from
    the Category model (categories table in the db).
  - Doing this allows for simpler querying of all categories at
    one time.

In order to separate the querying code for all values of either of the
Single Table Inheritance models, there is the InheritedModelLoader
class. This class can be configured via subclassing to query all
entities from the base model type, separate and sort them, and store
them as attributes on itself.

This commit also adds the StaticDataMixin class which allows loading
data into the db from a toml file. This is how all the characteristics
and categories are entered. This allows for an easy way to initially
store the data while also being easily repeatable and editable.

Another addition is the ProxyConversionMixin class which will convert a
proxy object into a SQLAlchemy model object from the db. This is needed
because the parser does not know about db objects. Yet, it is parsing
information that the StoredSpecial treats as a relationship. So rather
than explicitly setting the foreign key columns (which results in some
difficult to manage behavior when comparing parsed specials to stored
specials), we just use a 'ProxyAttribute'. This is then checked for when
performing various checks (like updating) and if a ProxyAttribute is
stored on a ParsedSpecial attribute, it is converted into the db object
it represents. This can be done efficiently because it is performed with
a 'get' query which will first search the identity map to see if the
object has already been queried. This will avoid repeated db queries for
objects that have already been loaded into memory.

This also adds the User model to the db. This is needed to protect the
endpoints of DVCTracker, which will now be more easily accessible
through the mobile app. The User model also adds the important_criteria
attribute, which is the new place that important_criteria is stored. It
is stored as a custom JSON type and that will convert any non python
objects into python objects when loading from the db and converts
python types that can't be converted with the JSON encoder, into JSON
supported representations when storing in the db.

And finally there were some changes made to the StoredSpecial model:
- The duration and price_per_night properties were changed to
  hybrid_property. This means they can now be used in queries.
- The class func 'get_columns' was replaced with 'get_core_keys'. This
  was needed because the way StoredSpecials are compared required
  reworking after adding in relationships.
- Storing old keys also needed to be improved. For regular attributes
  nothing fancy is required, but for hybrid_property attributes we now
  check what underlying attributes they are dependent on and set the old
  value for both the hybrid_property and the attributes it's dependent
  on.