Adds view for criteria form
This commit will add the necessary changes to introduce a view that has
a form that is used to interface with the important criteria.

This adds the CategoryChoices class, which will convert a list of
categories into the nested tuples of dictionaries that represent the
group layout of the choices for the MultiCheckboxField. A
CategoryChoices object can be configured through subclassing if
desired.

This change also moves the check to see whether a special is important
into the view function. This used to be done in the template. This is
more efficient because if the template wants to ask if the special is
important in multiple places it is easier if that info is passed in as
a bool, rather than having to either check repeatedly, or store the
value in a Jinja variable. It is also probably just more appropriate to
put this logic in the view function instead of in the template.