Updated confirmed parser with parser_map
The main change with this commit is to update the
dvcrentalstore_confirmed_2021 parser to now use the parser_map.toml
file. This mapping file maps the data the parser retrieves from the
website with the respective id from the db. When returning the id, the
id is returned not as itself, but in a ProxyAttribute object. This
allows the StoredSpecial to know that the object is supposed to
represent an object from the db.

This change also adds 'view' as a potential attribute on the
base_parser.