Update email templates
This updates the email templates to use a new version. This currently
only has a template for preconfirms but could accept discounted points
in the future if desired. Along with the new templates their are some
images and a slight change to the default date format for the
'datetimeformat' custom filter.

This also adds some more functionality to the StoredSpecial model:
- hybrid_property - 'price_per_point'
- properties - 'price_increased', 'price_per_night_increased', and
  'price_per_point_increased' which all return booleans if their was a
  change to their respective price attribute

These are used in the new templates.