Serve static content with 1 year cache expiration
This will serve all static content with an cache expiration of 1 year.

The reason behind this change is to try to reduce (to almost never)
the number of times the glyphs need to be requested. Since the project
is currently being used on a Heroku hobby tier the web dyno goes to
sleep after a period of inactivity. This is great because the main
function of this project is just running the 'update-specials' command
on a periodic basis. Running that command does not require much dyno
time and doesn't require the web dyno at all. This keeps the number of
free dyno hours consumed by this project very low. However, if web
requests were made each time an email was opened, that would wake the
dyno up much more frequently and consume, I would think, a significant
number of free hours. On top of that, there would be a noticeable delay
everytime the images were being loaded while the dyno wakes up. Neither
of these are ideal.

It is my understanding that many email clients (Gmail and iOS Mail in
particular) will cache images. I'm hoping that this is true and I'm
hoping that they respect the cache limit of 1 year that I have set on
these images. If they do, the glyphs will need to only be loaded the
first time an email is read. Each time thereafter the cached images will
be loaded.

In doing this I am also moving the bootstrap files into a folder that
specifies the version. This way, if I want to change bootstrap in the
future I can update the version directory and not have to worry about
the old version having been cached.