Improve url generation of static assets
Using 'url_for' and having 'SERVER_NAME' set as a config variable was
causing problems when trying to run flask locally. The 'SERVER_NAME'
needed to be set to one of the remote servers in order to generate the
correct urls for the images in the templates, but this caused problems
trying to serve webpages in a web browser. This commit introduces the
'static_url' function that can be used in templates to generate the
correct url for both web and email templates. It also works fine locally
and remotely since it uses a custom config variable 'EMAIL_SERVER_NAME',
which does not cause any problems serving webpages.