Update the email template and use Premailer
This change updates the email template while also no longer having the
css inlined in the stored template. This was becoming a challenge when
trying to update the template. For small changes it was manageable but
for anything where a lot of html was moving around it was difficult to
make sure the output from Postdrop was correctly converted into the
inlined template.

To address this, Premailer is now used. Premailer inlines the html
template if it is sending that html as an email. This makes it much
simpler to iterate on the email designs. This also eliminates the need
to have 'unsafe-inline' in the content security policy because when
viewing the specials through the web, we don't inline the css and the
stylesheet link can be used instead.