Added a logo and tweaked css styles
With this commit DVC Tracker gets a logo. There also were some css
changes relating to spacing (mostly around the logo) and colors. Now
that there is a logo, I feel I have a better understanding of the color
palette I would like to use through DVC Tracker.

The templates were also reorganized and the criteria_template was
refactored out into a base template and criteria template. This was done
so the upcoming User view can share the skeleton of its template with
criteria.