Add support for test view and email
This commit formally adds a test view and test cli command that sends a
test email.

It also refactors the special views and puts them in a child blueprint
of the main blueprint. Now the criteria view and any subsequent views
that do not show specials will not have the '/specials' prefix in their
url.