Add support for contacts linked to specific users
With this commit, email addresses, phone numbers, and APN tokens are now
connected to a specific user. It is also possible to view and update all
contact info directly through a webpage. I think this is the last big
change required to fully support multiple users.

There were a bunch of changes needed for this commit including:

- Updated db models to represent the various contacts. Email, Phone, and
  APN, are all single table inheritance models from Contact.
- New forms to process email addresses and phone numbers (and lists of
  these as well)
- With the new forms came some new fields and validators. And with the
  number of wtforms related classes growing, all were refactored into
  their own module.
- All of the notification methods now support the concept of 'users',
  rather than sending a message to all entities, it just will be sent
  to the entities related to the user.
- Additional views and templates that allow new contacts to be entered
  in a user friendly way.
- With the addition of the new views also came a refactoring of most of
  the prior view handling code.
- With the new templates...well more refactoring. This made it easier to
  ensure a consistent look across web pages.

There are also a few bugfixes sprinkled in as well.