Store the time the specials are updated
This stores the time the specials are updated in the Status model.
Before, the update time was just rendered into the special template
using 'datetime.now'. For emails this is okay, since they are created
when the specials are updated. But for the views, this time has less
meaning as it is just the current time. Now we can render the update
time for both emails and views. (Note: Because of this, it was necessary
to use dateutil for timezone handling while rendering the datetime.)

This also introduced the DefaultEntityMixin which allows easy access to
the 'default' db entity for a model. In this case, there is only one
Status entity that is needed. With this mixin, that entity can be easily
accesses with 'Status.default'.