Add 'test send' commands for text/apn
This adds commands to send a test for texts and apns. In order to
facilitate this, the respective 'send_update_{}' functions can now take
an optional message argument.

This also adds the 'cli' group into the cli module. This makes it easier
to add commands and then subsequently add them into Flask's cli group.