Fix when normalized data is returned
Before, normalized data was not returned when the *field* had any errors
on it. Now, normalized data will always be returned when it can, but we
won't include it in 'to_json' if there are errors present on the *form*.
When the form has errors, the appropriate original data is returned
instead.