Update dvcrentalstore parser_map
This updated the parser map to include 'Grand Villa - Wheelchair
Accessible'.

After looking at the DVC info a little more it seems the only resort
that has Wheelchair Accessibility listed is the Grand Californian. I
think it would just be better to treat the Wheelchair Accessible rooms
as regular rooms in order to avoid creating an extra room for each type.
It also seems somewhat misleading to include them as a parameter, since
only the Grand Californian offers them anyway.