Add GH Pages static assets release script
The ghp_release.py script, takes the static folder and the README and
copies them into the 'gh_pages_source' directory. This directory is not
tracked by git because its purpose is to stage the files and folders
that are supposed to be used by Github Pages. Inside this folder we have
a copy of the repo that is cloned from Github but only includes the
'gh-pages-static'  branch. In this repo we set the Github remote name to
'github-static' as another way of ensuring we only do this on this
special copy of the repo and not the normal repo on the 'master' branch.
After copying the static assets and README, the script adds all changes
to the index, commits them, squashes all commits, and pushes to Github
to the 'gh-pages-static' branch. The squash all step is used to get rid
of any previous versions of this branch. We do not want to hold onto the
past versions in this branch because the purpose of this is just for
building Github Pages with the current assets. All assets will also be
present  in the 'master' branch, along with any past versions.