Improve 'static_url' function and use more widely
This will allow greater control over how urls are created from
'static_url'. This includes the option to always generate a url that
uses the static server (rather than only when not in a request context)
and also adds the ability to specify a 'script_name' in the event the
assets are in a subdirectory on the static server. This also adds the
ability to include anchor links in the urls, which 'url_for' supports,
and is needed for the svg asset.

*HOWEVER*, it is important to note that you cannot load svgs from an
external server. So unfortunately for the time being, the svg asset will
need to continue to be fetched from the app directly. This is because
svgs don't currently support crossorigin. More info on that here:
https://oreillymedia.github.io/Using_SVG/extras/ch10-cors.html

This shouldn't be too bad from a UX perspective because the svgs are
only used on 'web only' views (since they aren't supported by email
clients). So the server will already be up and running when they are
requested and there won't be a noticeable delay in their loading. Once
they are loaded they also have a very long cache duration, so they
shouldn't need to be requested frequently.