Make special_type optional
This commit makes 'special_type' from ImportantCriteriaForm an optional
field. This was needed because validation was failing if all criteria
were removed from the criteria view. This is because there is a
'min_entries' of 1 set on the FieldList. So when submitting the form
with no entries, a blank one is created. But when validating this blank
entry, the 'special_type' has no raw data and will fail validation. By
making it optional, it doesn't matter if it doesn't include any raw
data. And since there is a default value for the field, any code that
relies on this always having a value, will still work as it will just be
filled with the default.