Install all dependencies into Virtual Environments
Before this I was installing everything at the system level because I
thought that was the way to go...since this is a container why bother
with the virtual environment? But after some research, it would seem it
is still good practice to use virtual environments. Now we install
pipenv into its own virtual environment and let it create another
virtual environment in the project directory to install the project's
dependencies into.