Recycle the database connection every hour
I have noticed a problem with the db connection on the CapRover server.
It seems to be closing the connection after a certain amount of time.
I'm going to try this approach, which should just recycle the connection
every hour, and hopefully this way the connection is always good when a
request is made.

This also updates the 'STRICT_SECURITY' config variable.