Fix test specials
With the updates to SQLAlchemy, it is now warning that the object
created in test is not attached to a session but was added to the
different characteristic relationships. Because it is added in the
relationships, SQLAlchemy would normally commit it to the db, but as it
is not in a session, it won't commit it. We don't want it to commit the
test object, so the warning is unnecessary/potentially confusing. To
avoid this we don't autoflush, and for the test view, we rollback the
session after rendering the html template.