Don't autoflush in mixins
Don't autoflush while querying inside the ProxyConversionMixin or the
DefaultEntityMixin. This is the specific cause of the new warnings
about a model object not being in a session but in a relationship. This
should fix the warning. It also means the prior no_autoflush context
that was added can be removed.