Switch to argon2 hashing algorithm for passwords
There is a problem with using bcrypt with passlib. This is passlib's
issue but passlib has not been updated in quite a while. Switching to
the argon2 algorithm avoids this problem and is a good option to use
anyway.

I was going to switch to use werkzeug's password handling functionality
but I like how passlib can automatically update old hashes. And it does
pack a lot of functionality in general. At the time of this commit it
seems like there is an effort to bring new maintenance back to the
project. I'll stick with it for now and keep an eye on the situation.

More info on maintaining the project:
https://foss.heptapod.net/python-libs/passlib/-/issues/187