Setup poetry in non-package mode
I'm not really using this as a package. And I think it would be better
to be explicit about that in Poetry. I've also removed the readme
and version items in the poetry tool section. I removed the readme
because it is causing problems with the check command. I removed the
version because I don't really want to do versions at this point, so for
now, I'm going to remove it.