Remove requirements.txt
I don't use the requirements.txt file, I haven't for a long time but
ketp it in the repo incase someone else might find use of it. I don't
believe anyone has, so it goes bye-bye. I also don't remember to keep it
synced with the lock file, so it isn't even accurate.